Fix and cleanup URL detection
Just checking for plainTextGet::startsWith is not enough. Just the
string "http" gets detected as a link, because it is valid URL (it's a
relative path to a file called "http") and it starts with http.

Now, we check, whether the URL scheme starts with http.

Also I cleaned up some of the checks and debug stuff that wasn't
necessary.