Make the settings migration actually work
It kept rewriting the format to the short one if you opted in to the
default one because the old settings variable was kept inside the
configuration file forever. This will delete it and further changes to
it won't be saved.

This still needs improvement though but it's a step in the right
direction