Implement a basic message search feature
It searches through all the lines that are currently fetched from the
server. Then you're able to list through them and scroll to them with
the buttons inside the buffer view.

There are no options now but it's case-insensitive and it looks only at
the nickname and the message body.