Fix handling of pointers for new buffer lines
WeeChat reuses pointers for objects that go out of scope. By default on my setup, 4096 lines per each buffer are kept and when the limit is reached, the resources get released and the pointer will be reused for a new line. This caused us issues with detecting "duplicate" lines in the buffer and some went missing occasionally. This fixes the problem.

Closes #138