Layout the QML modules as the engine expects it, do adjustments to make it work
Sorry for the vague commit message and I hope I'll be the only one reading git history in this repo because this one is huge and a bit messy. Basically, nothing has changed except it's not broken on iOS anymore - since the builds there are statically linked, some of the internal Qt modules were not installed into the build so it was not possible to use 1.4.7 at least. I really don't know why it didn't work with the previous layout but this seems to be a lot cleaner overall and it will contribute to further cleanups.

There's a few TODOs:
 - Move UI-related classes from Core to UI
 - Split Core into a Util library that's shared
 - Try making the modules dynamically linked so it's possible to use qHot