cmake: Prefix all custom variables
Generally, it is a good idea to prefix variables in CMake as subscripts can
accidentally use them. PROJECT_VERSION is a variable created by project()
call if you call with the VERSION argument, so it probably shouldn't be created
manually.

The specific reason for this change is to work around this bug:
https://bugreports.qt.io/browse/QTBUG-132421